requirements.txt
git+https://github.com/Rapptz/discord.py
asyncio
youtube_dl
PyNaCl
opuslib
discord.py[voice]